lock: fix so it can succeed after failing
Currently there is a bug where if the lock fails once (for example
because the file is locked) then it will never succeed, because
the map entry is never set to false in this case.

Rather than fix in every place, we factor out the common code
from all the implementations and apply the fix there.

We also fix the issue that if the file is locked and the portable code
is being used, the error message says "file has non-zero size"
rather than "already locked".

The tests are changed to catch this case - they fail as expected when run
against the old code.

Change-Id: Ie10eff1ff6a430da6f7ee56c76b0c792da337fcb