Added Solaris to the list of builds that are compatible with 'lock_unix'.
I think it might have been left out initially because fcntl() may have
been confused with flock(). The Go function used by 'lock_unix' is
called FcntlFlock(). This somewhat seems to imply that 'flock()' is
used. However the doc for the function states: "FcntlFlock performs a
fcntl syscall" so flock() is not involved.

Solaris doesn't support flock() but does support fcntl() (which is part
of POSIX). For more info on Solaris and locking see:
https://www.perkin.org.uk/posts/solaris-portability-flock.html

I've tested the lock in a SmartOS (Illumos / Solaris) zone and it worked
as expected.