jsonconfig: add test for numbers
The fact that JSON numbers decode as float64, even though it's mentioned
on the encoding/json doc, is confusing enough that people (myself
included), sometimes try to change the type assertions to float64 in
assertions to int or int64.

For example: https://review.gerrithub.io/#/c/346661/

This CL adds a decoding test for numbers, as an attempt to prevent that
confusion.

Change-Id: I26f4c0279cd77c02aef7e27f37793621c295c1d7